pictl: `commenttimestamps` improvements.
This diff fixes a bug in `pictl commenttimestamps` where no timestamps
were fetched if the no comment IDs were provided, this is not the
expected behavior as it says in the `commenttimestamps` command help
message:

> If comment IDs are not provided then the timestamps for all comments will be
returned.

In addition, this commit adds pagination to the `commenttimestamps`
command to allow fetching all comment timestamps page by page when
a proposal has more than one page of comments. This also allows users to
request the timestamps of more than one page using the `commentIDs`
command argument.