comments: Cache final timetamps.
This diff improves the performance of retrieving comment timestamps. It
uses the key-value store to cache final timestamps which have been
timestamped on the DCR chain as the performance of evaluating it on
runtime is proportional to the tree size which makes it unacceptably
slow in prod where the trees has ~30k leaves.