pi: Add `LegacyToken` to `ProposalMetadata`.
This diff adds a new `LegacyToken` field to the `ProposalMetadata` struct. It
will be populated only for legacy gitbe proposals which will be imported
to the backend in the next step.

This commit also adds a validation to the `pi` plugin to ensure that
the new `LegacyToken` field in not set during the normal proposal
submissions.