politeiawww: Add plugin model.
This commit lays the foundation for the politeia plugin model. From a
high level:

- The simplified plugin API defined in this commit will be the only API.
  Everything will be turned into a plugin. All existing APIs will be
  deprecated in the future and eventually removed.

- The politeiad API will be converted into a `records` plugin in the
  future.  politeiad and politeiawww will be combined into a single
  daemon.

- politeia can be run with the user layer enabled or disabled. Using the
  above mentioned `records` plugin as an example, when the user layer is
  disabled, anyone with access to the API can run any of the registered
  plugin commands.  This is similar to how the politeiad API currently
  functions. Running with the user layer enabled adds user authorization
  onto the execution of plugin commands.  This is similar to how the
  politeiawww API currently functions.  The user layer in enabled by
  default.

politeia can be broken into three layers.
- API layer
- Backend layer
- Plugin layer

#### API layer

The `api/http` package adds a generic HTTP API for plugin commands.
Additional APIs that use different communication protocols, such as a
websockets API, can be added in the future.

All other APIs will be deprecated and removed once the existing
functionality has been ported over to plugins and clients have migrated.

The plugin payloads and error handling were designed to be agnostic of
the communication protocol being used. The same request and response
payloads that are used in the HTTP API can also be used in a websockets
API or in any other additional communication methods that may be added
in the future.

#### Backend Layer

The backend layer handles payload encoding/decoding, session managment,
user managment, and acting as the intermediary between the API and the
plugins.

Plugins do not have direct access to the sessions database or to the
user database. The backend handles these connections and only provides
the plugins with data that they are allowed to access.  Plugins are able
to update session and user data during the execution of certain types of
commands, ex. plugin writes, and the backend will persist those changes
once the command has finished executing.

All operations performed during the execution of plugin write commands
are atomic.

#### Plugin Layer

The plugin layer contains the plugin implementations.

Plugins are only provided access to the data that they own. If they wish
to retrieve data from a separate plugin, they must use the standard
plugin API.  This enforces a clear separation of concerns.

Plugins have two ways to save user data:

1. Update the user object that is provided to the plugin during command
   execution. These updates are saved to the database by the backend on
   successful completion of the plugin command.  Plugins are able to
   specify whether data should be saved clear text or encrypted. The
   plugins do not actually have to encrypt or decrypt any data. This is
   all handled by the backend.

2. Plugins can create and manage a database table themselves to store
   plugin user data in. This option should be reserved for data that
   would cause performance issues if saved to the global user object or
   if the plugin needs to be able to query the data using SQL.

### HTTP API

The HTTP API consists of six routes.

`Version` - a GET request route that returns server and plugin version
information. The version route also sets CSRF tokens using the double
submit cookie method.

`Policy` - a GET request route that returns API policy information.

`NewUser` - a POST request that executes a plugin command that creates a
new user in the database. This was required to be broken out into a new
route because the plugin layer does not have access to user database
method for inserting new users.

`Write` - a POST request route that executes a plugin write command.

`Read` - a POST request route that executes a single read-only plugin
command.  This route is used for expensive plugin commands that are not
allowed to be executed as part of a batch.

`ReadBatch` - a POST request route that executes a batch of read-only
plugin commands. The default batch size is 20 plugin commands and can be
configured at runtime using the `--pluginbatchlimit` config setting.

#### CSRF Protection

CSRF tokens are set by the `Version` route using the double submit
cookie method. All routes that allow writes are CSRF protected. This
currently includes the `NewUser` route and the `Write` route. Clients
MUST make a successful call to the `Version` route before they'll be
able to access CSRF protected routes.