politeiad: Add legacypoliteia convert cmd.
The legacypoliteia tool is a CLI tool that will be used to convert and
import legacy git backend data into the tstore backend. The tool will
have two commands: convert and import.

The convert command converts the git backend data into tstore backend
data types and saves the converted data locally to JSON files. The
import command will import these JSON files into the tstore backend.

This commit adds the legacypoliteia convert command.

The convert command walks a politeia git repo and performs the
conversion on each proposal found in the repo. This is not a simple
conversion and some tstore types require retrieving data from external
APIs. Listed below are the performance bottlenecks of the tool and the
external API calls that are required.

- Parsing the git commit log for commit timestamps so that each cast
  vote can be associated with the timestamp of when it was committed to
  the ballots journal. Exact timestamps of when a vote was cast does not
  exist, so this is the closest substitute.

- Retrieving the largest commitment address for each vote from dcrdata.
  This address is used to verify the vote signature and is part of the
  ticketvote plugin vote structure, but was not part of the git backend
  vote structure.

- Retrieving the user IDs from the politeia API using the public key
  that was used to create signatures for various pieces of data
  (records, comments, etc). These user IDs were not saved as part of the
  git backend data structures, but are part of the tstore and plugin
  data structures.

All of the legacy data types can be found in the tool's gitbe package.
The convert.go file contains all of the data conversions in an easy to
audit format.

A proposal that was voted on takes <30 seconds to convert. A proposal
that was not voted on takes 1-2 seconds to convert. The legacy politeia
data can be fully converted in ~50 minutes. The tool uses <100 MiB of
memory during execution.