mysql: Fix InsertUser identity table bug.
This fixes a bug in the InsertUser function of the mysql user database
implementation.

When a new user record is inserted into the database, both the users
table and the identities table must be updated. The InsertUser function
was not updating the identities table, which was causing user by public
key lookups to fail.

The InsertUser function is only used during database migrations. This
bug does not impact regular politeia use.

This bug was introduce by 0b1982a.