Docs.
This commit rewrites the ticketvote inventory cache.

The initial plugin implementations were part of the extremely large
commit that added the tstore backend. The caches were saved to the file
system as a quick and dirty way to get things into production, with the
intention of revisting them at a later date. API was later added that
allowed the plugins to cache data in the tstore database.

The new inventory cache has an improved API, better documentation, and
was moved off of the file system and into the tstore database.