politeiad: Remove leveldb tstore config option.
This commit removes the ability to use leveldb as the tstore key-value
store.

tstore already requires a MySQL instance to be running since trillian
uses MySQL. It doesn't make a whole lot of sense to have a config option
to use leveldb as the tstore key-value store since MySQL is already a
requirement. Removing this option simplifies the politeiad setup and
makes the code easier to maintain. MySQL is now used by default.

If we move the trillian functionality into politeia and remove the
trillian dependency at some point in the future, then it may make sense
to add a leveldb key-value store option back in.