ticketvote: Add the timestamp to the vote summary.
This commit adds the timestamp of the vote status change to the vote
summary structure.

The function that builds the vote summary contains the logic for
determining the vote status of a record. Including the timestamp of the
vote status change in the vote summary ensures that the logic for
determing the vote status and it's associated timestamp remains in one
place.

This addition also allows the vote inventory to be rebuilt using just
the vote summaries, without having to pull any additional data.