ticketvote: Move runoff vote subs cache to db.
This commit rewrites the runoff vote submissions cache and moves it from
the file system to the tstore provided plugin cache.

The initial plugin implementations were part of an extremely large
commit that added the tstore backend. The caches were saved to the file
system as a quick and dirty way to get things into production, with the
intention of revisting them at a later date. API was later added that
allowed the plugins to cache data in the tstore database.