politeiawww: Remove legacy websockets.
This commit removes the websockets package and the legacy websocket
route handlers. The legacy websockets were never used for anything and
this code was the equivalent to dead code. If websockets are added in
the future, they'll be added as part of the plugin model and the
implementation will be quite different.