tstore: Improve blobs not found error.
This commit improves the error message that is returned when the
expected number of blobs are not found in the key-value store.

Previously, the keys that were not found were being included in the
error message. This made the error message illegible when a large number
of keys were being requested, like when all of a proposal's cast votes
are being requested.