legacypoliteia: Add import command.
This adds an `import` command to the legacypoliteia tool. The import
command imports the JSON data into the tstore backend that was saved to
disk during the execution of the convert command.

Using these two commands (convert and import), the legacy git backend
proposals can now be parsed from the politiad git repo, converted to
types supported by the tstore backend, and imported into the tstore
backend.

Execution of the import command takes ~7.5 hrs to complete when importing
the 115 mainnet Politeia proposals from the git repo.

The performance bottleneck for this command is the trillian log server
(tlog server). ~50 leaves/sec can be appended onto a tlog tree without
causing any issues. This means that importing 10,000 proposal votes will
take ~200 seconds (3 minutes, 20 seconds). The vast majority of the
execution time of this command is spent importing proposal votes.

The command is relatively light weight. It's memory footprint should
stay under 100 MiB and CPU usage should be minimal.

## Data Changes

One of the main issues with importing legacy proposal data into the
tstore backend is that the client and server signatures of the data are
broken. For this reason, all signatures have been removed from the
imported data. 

The `overwriteProposalFields` function in `legacypoliteia/proposal.go`
documents all changes that are being made to the legacy data and the
reason for each change.

Some legacy data that was the result of bugs in the legacy politeia
backend are omitted imported into the tstore backend. These include
duplicate comments and duplicate proposal votes.

The original data, signatures, timestamps for the legacy proposals can
be found in the git repo.

## Trillian Log Server Issue

One issue that was repeatedly encountered was a bug with the tlog server
when appending cast vote leaves onto a tlog tree. A small number of
requests will intermittently hang. tlog will not throw any errors or
give any indication that anything is wrong. tlog will sometimes require
a hard restart. Upon restart of tlog, the leaves that were being
appended during the bad request are usually present on the tlog tree.

This command was written to mitigate this issue as much as possible, but
tlog may require a hard restart during import. The command will get
stuck in a retry loop until the tlog server is reset.

## Running Locally

See the example below for testing this tool locally. politeiad and
politeiawww do not need to be running during the import. The trillian
log server, trillian log signer, and MySQL must running in order to
import the data.

Once the data has been imported, politeiad must be reset using the
--fsck flag so that the caches are rebuilt to include the newly imported
data.

Note, this is mainnet data and will be imported into the mainnet
databases. Make sure you are running politeiad and politeiawww against
mainnet, not testnet, when you turn them on.

```
# Clone the git repo for the mainnet politeia data

$ git clone https://github.com/decred-proposals/mainnet.git 

# Convert the legacy git backend data to tstore backend types.
# The following directory will be created during the execution
# of this command and the output will be saved to it as JSON
# formatted data.
#
# Output directory: ./legacy-politeia-data

$ legacypoliteia convert ./mainnet

# Import the converted data into the tstore backend.
#
# The --stubusers flag is used when testing locally. This
# will create stubs in the user database for all of the user
# IDs that are found in the legacy proposal data. This will
# allow you to run politeiawww without any 'user not found'
# errors when retrieving legacy data.

$ legacypoliteia import ./legacy-politeia-data --stubusers
```