dbutil: Improve migrate cmd.
This commit updates the politeiawww_dbutil command to only read a single
user into memory at a time. It was previously reading all users into
memory and then performing the migration. It also adds in some
additional validation to ensure that an incorrect paywall address index
is not being migrated over.