dbutil: Update dump cmd.
This commit updates the dump command so that it works with all of the
user database implemenations. It also simplifies the implementation to
only allow for dumping a single user to stdout. There is no use case
that requires dumping all of the users at once. The migrate command
should be used when migrating databases, not the dump command.