plugins/pi: Fix author update ordering bug.
This commit fixes a bug in a function that was used to determine the
most recent author update. The function was using the wrong timestamp to
determine the order of the author updates. It was using the timestamp
of the most recent comment edit when it should have been using the
timestamp of the comment creation.

This bug was introduced by b43eebd.