pivoter: Consolidate cast vote creation.
This consolidates the logic that creates the signed votes to cast by
creating them without the signature first and then obtaining the
signatures from the wallet and setting those signature into the existing
unsigned votes.

This approach is simpler and less error prone than the current one that
passes separate arrays of eligible tickets, signatures, and associated
vote details around.